Code refactor - add world class
